+++ warp/native/warp.h
@@ -100,7 +100,7 @@ extern "C"
     WP_API void* cuda_load_module(const char* ptx);
     WP_API void cuda_unload_module(void* module);
     WP_API void* cuda_get_kernel(void* module, const char* name);
-    WP_API size_t cuda_launch_kernel(void* kernel, int dim, void** args);
+    WP_API size_t cuda_launch_kernel(void* kernel, size_t dim, void** args);
 
 } // extern "C"
